requirements.txt
fastapi
fastapi-cli
httpx
uvicorn[standard]
python-dotenv
jinja2
itsdangerous
